significant improvements to ip address handling and now supports ipv6 throughout.
git-svn-id: https://plugins.svn.wordpress.org/wp-simple-firewall/trunk@1007347 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ src/icwp-processor-basedb.php
@@ -90,6 +90,7 @@ abstract class ICWP_WPSF_BaseDbProcessor extends ICWP_WPSF_Processor_Base {
 	 */
 	protected function initializeTable() {
 		if ( $this->getTableExists() ) {
+			$this->recreateTable();
 			$sFullHookName = $this->getDbCleanupHookName();
 			add_action( $sFullHookName, array( $this, 'cleanupDatabase' ) );
 		}
@@ -191,8 +192,12 @@ abstract class ICWP_WPSF_BaseDbProcessor extends ICWP_WPSF_Processor_Base {
 	 * Will recreate the whole table
 	 */
 	public function recreateTable() {
-		$this->dropTable();
-		$this->createTable();
+		if ( $this->getFeatureOptions()->getOpt( 'recreate_database_table', false ) ) {
+			$this->getFeatureOptions()->setOpt( 'recreate_database_table', false );
+			$this->getFeatureOptions()->savePluginOptions();
+			$this->dropTable();
+			$this->createTable();
+		}
 	}
 	
 	/**
